+++ components/zcash_protocol/src/constants.rs
@@ -42,8 +42,10 @@ pub const V6_TX_VERSION: u32 = 6;
 pub const V6_VERSION_GROUP_ID: u32 = 0xD884B698;
 
 /// Transaction version 7, introduced by the NuTachyon network upgrade.
+#[cfg(zcash_unstable = "nutachyon")]
 pub const V7_TX_VERSION: u32 = 7;
 /// The placeholder version group ID for Zcash v7 transactions ("tach" in ASCII).
+#[cfg(zcash_unstable = "nutachyon")]
 pub const V7_VERSION_GROUP_ID: u32 = 0x74616368;
 
 /// The maximum size in bytes of a Zcash block, and therefore the maximum size of any single