+++ components/zcash_protocol/src/local_consensus.rs
@@ -33,6 +33,7 @@ use crate::consensus::{BlockHeight, NetworkType, NetworkUpgrade, Parameters};
 ///         nu6_3: Some(BlockHeight::from_u32(1)),
 ///         #[cfg(zcash_unstable = "nu7")]
 ///         nu7: Some(BlockHeight::from_u32(1)),
+///         #[cfg(zcash_unstable = "nutachyon")]
 ///         nu_tachyon: Some(BlockHeight::from_u32(1)),
 ///     };
 ///     ```
@@ -51,6 +52,7 @@ pub struct LocalNetwork {
     pub nu6_3: Option<BlockHeight>,
     #[cfg(zcash_unstable = "nu7")]
     pub nu7: Option<BlockHeight>,
+    #[cfg(zcash_unstable = "nutachyon")]
     pub nu_tachyon: Option<BlockHeight>,
 }
 
@@ -74,6 +76,7 @@ impl Parameters for LocalNetwork {
             NetworkUpgrade::Nu6_3 => self.nu6_3,
             #[cfg(zcash_unstable = "nu7")]
             NetworkUpgrade::Nu7 => self.nu7,
+            #[cfg(zcash_unstable = "nutachyon")]
             NetworkUpgrade::NuTachyon => self.nu_tachyon,
         }
     }
@@ -101,6 +104,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        #[cfg(zcash_unstable = "nutachyon")]
         let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
@@ -116,6 +120,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            #[cfg(zcash_unstable = "nutachyon")]
             nu_tachyon: Some(expected_nu_tachyon),
         };
 
@@ -132,8 +137,11 @@ mod tests {
         // nu7 must not be activated at or below the nu6_3 height
         #[cfg(zcash_unstable = "nu7")]
         assert!(!regtest.is_nu_active(NetworkUpgrade::Nu7, expected_nu6_3));
-        assert!(!regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu6_3));
-        assert!(regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu_tachyon));
+        #[cfg(zcash_unstable = "nutachyon")]
+        {
+            assert!(!regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu6_3));
+            assert!(regtest.is_nu_active(NetworkUpgrade::NuTachyon, expected_nu_tachyon));
+        }
     }
 
     #[test]
@@ -150,6 +158,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        #[cfg(zcash_unstable = "nutachyon")]
         let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
@@ -165,6 +174,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            #[cfg(zcash_unstable = "nutachyon")]
             nu_tachyon: Some(expected_nu_tachyon),
         };
 
@@ -213,6 +223,7 @@ mod tests {
             regtest.activation_height(NetworkUpgrade::Nu7),
             Some(expected_nu7)
         );
+        #[cfg(zcash_unstable = "nutachyon")]
         assert_eq!(
             regtest.activation_height(NetworkUpgrade::NuTachyon),
             Some(expected_nu_tachyon)
@@ -233,6 +244,7 @@ mod tests {
         let expected_nu6_3 = BlockHeight::from_u32(10);
         #[cfg(zcash_unstable = "nu7")]
         let expected_nu7 = BlockHeight::from_u32(11);
+        #[cfg(zcash_unstable = "nutachyon")]
         let expected_nu_tachyon = BlockHeight::from_u32(12);
 
         let regtest = LocalNetwork {
@@ -248,6 +260,7 @@ mod tests {
             nu6_3: Some(expected_nu6_3),
             #[cfg(zcash_unstable = "nu7")]
             nu7: Some(expected_nu7),
+            #[cfg(zcash_unstable = "nutachyon")]
             nu_tachyon: Some(expected_nu_tachyon),
         };