+++ zcash_client_backend/src/data_api/testing.rs
@@ -1778,6 +1778,7 @@ impl TestBuilder<(), ()> {
         nu6_3: None,
         #[cfg(zcash_unstable = "nu7")]
         nu7: None,
+        #[cfg(zcash_unstable = "nutachyon")]
         nu_tachyon: None,
     };