+++ zcash_primitives/src/transaction/builder.rs
@@ -904,6 +904,7 @@ impl<P, U> Builder<P, U> {
                 BranchId::Nu6_3 => true,
                 #[cfg(zcash_unstable = "nu7")]
                 BranchId::Nu7 => true,
+                #[cfg(zcash_unstable = "nutachyon")]
                 BranchId::NuTachyon => true,
                 _ => false,
             };
@@ -1961,6 +1962,7 @@ mod tests {
             nu6_3: Some(BlockHeight::from_u32(10)),
             #[cfg(zcash_unstable = "nu7")]
             nu7: None,
+            #[cfg(zcash_unstable = "nutachyon")]
             nu_tachyon: None,
         }
     }
@@ -1979,6 +1981,7 @@ mod tests {
             nu6_2: Some(BlockHeight::from_u32(9)),
             nu6_3: Some(BlockHeight::from_u32(10)),
             nu7: Some(BlockHeight::from_u32(11)),
+            #[cfg(zcash_unstable = "nutachyon")]
             nu_tachyon: None,
         }
     }