+++ zcash_primitives/src/transaction/sighash.rs
@@ -50,6 +50,8 @@ pub fn signature_hash<
 
         TxVersion::V5 => v5_signature_hash(tx, signable_input, txid_parts),
 
-        TxVersion::V6 | TxVersion::V7 => v6_signature_hash(tx, signable_input, txid_parts),
+        TxVersion::V6 => v6_signature_hash(tx, signable_input, txid_parts),
+        #[cfg(zcash_unstable = "nutachyon")]
+        TxVersion::V7 => v6_signature_hash(tx, signable_input, txid_parts),
     })
 }